+++ src/components/button-show.js
@@ -1,0 +1,7 @@
+const createButtonShowMoreFilmsTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createButtonShowMoreFilmsTemplate};
