+++ src/components/films-container.js
@@ -1,0 +1,7 @@
+const createFilmsContainerTemplate = () => {
+  return (
+    `<div class="films-list__container"></div>`
+  );
+};
+
+export {createFilmsContainerTemplate};
